refactored local opt classes, chromosome and generator accept list of values for optimization
