+++ bingo/MultipleFloatChromosome.py
@@ -1,11 +1,12 @@
 from .Base.ContinuousLocalOptimization import ChromosomeInterface
+from .Util.ArgumentValidation import argument_validation
 from .MultipleValues import MultipleValueChromosome, MultipleValueGenerator
 
-class MultiValueContinuousLocalOptimization(MultipleValueChromosome, ChromosomeInterface):
+class MultipleFloatChromosome(MultipleValueChromosome, ChromosomeInterface):
 
-    def __init__(self, list_of_values, needs_opt_list):
+    def __init__(self, list_of_values, needs_opt_list=[]):
         super().__init__(list_of_values)
-        self.needs_opt_list = needs_opt_list
+        self._needs_opt_list = needs_opt_list
 
     def needs_local_optimization(self):
         """Does the individual need local optimization
@@ -15,8 +16,9 @@ class MultiValueContinuousLocalOptimization(MultipleValueChromosome, ChromosomeI
         bool
             Individual needs optimization
         """
+        if not self._needs_opt_list:
+            return False
         return True
-
 
     def get_number_local_optimization_params(self):
         """Get number of parameters in local optimization
@@ -26,7 +28,7 @@ class MultiValueContinuousLocalOptimization(MultipleValueChromosome, ChromosomeI
         int
             number of paramneters to be optimized
         """
-        return len(self.needs_opt_list)
+        return len(self._needs_opt_list)
 
     def set_local_optimization_params(self, params):
         """Set local optimization parameters
@@ -36,10 +38,26 @@ class MultiValueContinuousLocalOptimization(MultipleValueChromosome, ChromosomeI
         params : list-like of numeric
                  Values to set the parameters
         """
-        for i, index in enumerate(self.needs_opt_list):
+        for i, index in enumerate(self._needs_opt_list):
             self.list_of_values[index] = params[i]
 
-class MultiValueContinuousLocalOptimizationGenerator(MultipleValueGenerator):
+class MultipleFloatChromosomeGenerator(MultipleValueGenerator):
+    """Generation of a population of Multi-Value Chromosomes
+
+    Parameters
+    ----------
+    random_value_function : user defined function
+        A function that returns a list of randomly generated values.
+        This list is then passed to the ``MultipleValueChromosome``
+        constructor.
+    values_per_chromosome : int
+        The number of values that each chromosome will hold
+    """
+    @argument_validation(values_per_chromosome={">=": 0})
+    def __init__(self, random_value_function, values_per_chromosome,
+                 needs_opt_list=[]):
+        super().__init__(random_value_function, values_per_chromosome)
+        self._needs_opt_list = needs_opt_list
 
     def __call__(self):
         """Generation of a population of size 'population_size'
@@ -52,5 +70,4 @@ class MultiValueContinuousLocalOptimizationGenerator(MultipleValueGenerator):
         out : a MultipleValueChromosome
         """
         random_list = self._generate_list(self._values_per_chromosome)
-        needs_op_list = [1, 4, 5]
-        return MultiValueContinuousLocalOptimization(random_list, needs_op_list)
+        return MultipleFloatChromosome(random_list, self._needs_opt_list)
